+++ sources/gmtime.c
@@ -12,7 +12,7 @@
 #define is_leap(y)  ((y) % 4 == 0)
 
 
-const long _month_days[] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
+const int _month_days[] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
 
 enum { January, February, March, April, May, June, July, August, September, October, November, December, Months };
 enum { Sunday, Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Weekdays };
@@ -26,8 +26,8 @@ gmtime(const time_t* timep)
     time_t t   = *timep;
     int    neg = (t < 0);
     int    step;
-    int    days;
-    int    seconds;
+    long   days;
+    long   seconds;
     int    leap;
     int    monthdays;
 
@@ -38,13 +38,13 @@ gmtime(const time_t* timep)
         step = 1;
     }
 
-    days    = t / 86400;
-    seconds = t % 86400;
+    days    = t / 86400L;
+    seconds = t % 86400L;
 
     if (neg) {
         tm.tm_wday = (Wednesday + days) % Weekdays;
         tm.tm_year = 1969;
-        seconds    = 86399 - seconds;
+        seconds    = 86399L - seconds;
     } else {
         tm.tm_wday = (Thursday + days) % Weekdays;
         tm.tm_year = 1970;
@@ -96,9 +96,9 @@ gmtime(const time_t* timep)
 
     tm.tm_year -= 1900;
     tm.tm_mday += step * days;
-    tm.tm_hour  = seconds / 3600;
-    tm.tm_min   = (seconds % 3600) / 60;
-    tm.tm_sec   = seconds % 60;
+    tm.tm_hour  = (int)(seconds / 3600L);
+    tm.tm_min   = (int)((seconds % 3600L) / 60L);
+    tm.tm_sec   = (int)(seconds % 60L);
     tm.tm_isdst = -1;
 
     return &tm;
